+++ client/components/book-views/book-component.js
