+++ client/components/index.js
@@ -5,15 +5,15 @@
  */
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
-export {default as SingleBook} from './single-book'
-export {default as AllBooks} from './all-books-view'
-export {default as BooksByGenre} from './books-by-genre-view'
+export {default as SingleBook} from './book-views/single-book'
+export {default as AllBooks} from './book-views/all-books-view'
+export {default as BooksByGenre} from './book-views/books-by-genre-view'
 export {Login, Signup} from './auth-form'
 export {default as LandingPage} from './landing-page'
-export {default as Cart} from './cart'
+export {default as Cart} from './payment/cart'
 export {default as Checkout} from './payment/checkout'
 export {default as StripeCheckout} from './payment/stripe-checkout'
 export {default as ReviewOrder} from './payment/review-order'
-export {default as AllStaffs} from './all-staff'
-export {default as SingleStaff} from './single-staff'
+export {default as AllStaffs} from './staff-views/all-staff'
+export {default as SingleStaff} from './staff-views/single-staff'
 export {default as Footer} from './footer'
