+++ server/api/Orders.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const {Order} = require('../db/models/Index')
+const {Order} = require('../db/models')
 module.exports = router
 
 router.get('/', async (req, res, next) => {
