+++ server/api/Users.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const {User} = require('../db/models/Index')
+const {User} = require('../db/models')
 module.exports = router
 
 router.get('/', async (req, res, next) => {
