+++ server/db/models/test-specs/Book.spec.js
@@ -1,7 +1,7 @@
 // Assertions
 const chai = require('chai')
 const expect = chai.expect
-const db = require('../Index')
+const db = require('../../index')
 
 // Models
 const Book = db.model('book')
