+++ server/db/models/test-specs/Genre.spec.js
@@ -1,5 +1,5 @@
 const {expect} = require('chai')
-const db = require('../Index')
+const db = require('../../index')
 
 const Genre = db.model('genre')
