+++ server/db/models/test-specs/User.spec.js
@@ -1,7 +1,7 @@
 /* global describe beforeEach it */
 
 const {expect} = require('chai')
-const db = require('../Index')
+const db = require('../../index')
 const User = db.model('user')
 
 describe('User model', () => {
